myproject/requirements.txt
django==4.2.13
uwsgi==2.0.26
aws-opentelemetry-distro==0.3.0